requirements.txt
httpx
sentence-transformers
ffmpeg
audio2numpy

pydub
ffmpeg
requests